get back original RRDSI
