+++ deeplearning4j-core/src/main/java/org/deeplearning4j/datasets/datavec/RecordReaderDataSetIterator.java
@@ -18,7 +18,8 @@
 
 package org.deeplearning4j.datasets.datavec;
 
-import lombok.*;
+import lombok.Getter;
+import lombok.Setter;
 import lombok.extern.slf4j.Slf4j;
 import org.datavec.api.io.WritableConverter;
 import org.datavec.api.io.converters.SelfWritableConverter;
@@ -29,11 +30,7 @@ import org.datavec.api.records.reader.RecordReader;
 import org.datavec.api.records.reader.SequenceRecordReader;
 import org.datavec.api.writable.Writable;
 import org.datavec.common.data.NDArrayWritable;
-import org.deeplearning4j.exception.DL4JInvalidConfigException;
 import org.deeplearning4j.exception.DL4JInvalidInputException;
-import org.nd4j.linalg.api.memory.MemoryWorkspace;
-import org.nd4j.linalg.api.memory.conf.WorkspaceConfiguration;
-import org.nd4j.linalg.api.memory.enums.*;
 import org.nd4j.linalg.api.ndarray.INDArray;
 import org.nd4j.linalg.dataset.DataSet;
 import org.nd4j.linalg.dataset.api.DataSetPreProcessor;
@@ -42,19 +39,16 @@ import org.nd4j.linalg.factory.Nd4j;
 import org.nd4j.linalg.util.FeatureUtil;
 
 import java.io.IOException;
-import java.util.*;
-import java.util.concurrent.*;
-import java.util.concurrent.atomic.AtomicBoolean;
-import java.util.concurrent.atomic.AtomicInteger;
-import java.util.concurrent.atomic.AtomicLong;
-import java.util.concurrent.locks.LockSupport;
+import java.util.ArrayList;
+import java.util.Collections;
+import java.util.Iterator;
+import java.util.List;
 
 
 /**
  * Record reader dataset iterator
  *
  * @author Adam Gibson
- * @author raver119@gmail.com
  */
 @Slf4j
 public class RecordReaderDataSetIterator implements DataSetIterator {
@@ -72,21 +66,10 @@ public class RecordReaderDataSetIterator implements DataSetIterator {
     protected boolean regression = false;
     @Getter
     protected DataSetPreProcessor preProcessor;
-    private AtomicBoolean collectMetaData = new AtomicBoolean(false);
-    private volatile int prefetchSize = 8;
 
-    private BlockingQueue<Future<DataSet>> buffer;
-    private Future<DataSet> nextElement = null;
-    private Future<DataSet> terminator = new DummyFuture();
-    private ThreadPoolExecutor executor;
-    private AsyncPrefetchThread thread;
-    private final String guid = java.util.UUID.randomUUID().toString();
-    private AtomicBoolean wasTriggered = new AtomicBoolean(false);
-    private boolean useWorkspaces = true;
-    private AtomicLong counterGlobal = new AtomicLong(0);
-    private int workers = 2;
-
-    // RecordReaderDataSetIterator(recordReader, AppConfig.batchSize, 1, AppConfig.numLabels);
+    @Getter
+    @Setter
+    private boolean collectMetaData = false;
 
     public RecordReaderDataSetIterator(RecordReader recordReader, WritableConverter converter, int batchSize) {
         this(recordReader, converter, batchSize, -1,
@@ -96,11 +79,6 @@ public class RecordReaderDataSetIterator implements DataSetIterator {
     public RecordReaderDataSetIterator(RecordReader recordReader, int batchSize) {
         this(recordReader, new SelfWritableConverter(), batchSize, -1,
                 recordReader.getLabels() == null ? -1 : recordReader.getLabels().size());
-    }
-
-    public RecordReaderDataSetIterator(RecordReader recordReader, int batchSize, int numThreads) {
-        this(recordReader, new SelfWritableConverter(), batchSize, -1,
-                recordReader.getLabels() == null ? -1 : recordReader.getLabels().size(), numThreads);
     }
 
     /**
@@ -125,11 +103,6 @@ public class RecordReaderDataSetIterator implements DataSetIterator {
     public RecordReaderDataSetIterator(RecordReader recordReader, WritableConverter converter, int batchSize,
                                        int labelIndex, int numPossibleLabels) {
         this(recordReader, converter, batchSize, labelIndex, numPossibleLabels, -1, false);
-    }
-
-    public RecordReaderDataSetIterator(RecordReader recordReader, WritableConverter converter, int batchSize,
-                                       int labelIndex, int numPossibleLabels, int numThreads) {
-        this(recordReader, converter, batchSize, labelIndex, numPossibleLabels, -1, false, numThreads);
     }
 
     public RecordReaderDataSetIterator(RecordReader recordReader, int batchSize, int labelIndex, int numPossibleLabels,
@@ -157,11 +130,6 @@ public class RecordReaderDataSetIterator implements DataSetIterator {
         this(recordReader, converter, batchSize, labelIndex, labelIndex, numPossibleLabels, maxNumBatches, regression);
     }
 
-    public RecordReaderDataSetIterator(RecordReader recordReader, WritableConverter converter, int batchSize,
-                                       int labelIndex, int numPossibleLabels, int maxNumBatches, boolean regression, int numThreads) {
-        this(recordReader, converter, batchSize, labelIndex, labelIndex, numPossibleLabels, maxNumBatches, regression, numThreads, 8, false, true);
-    }
-
 
     /**
      * Main constructor
@@ -177,23 +145,6 @@ public class RecordReaderDataSetIterator implements DataSetIterator {
     public RecordReaderDataSetIterator(RecordReader recordReader, WritableConverter converter, int batchSize,
                                        int labelIndexFrom, int labelIndexTo, int numPossibleLabels, int maxNumBatches,
                                        boolean regression) {
-        this(recordReader, converter, batchSize, labelIndexFrom, labelIndexTo, numPossibleLabels, maxNumBatches, regression, 1, 8, false, true);
-
-    }
-
-    public RecordReaderDataSetIterator(RecordReader recordReader, WritableConverter converter, int batchSize,
-                                       int labelIndexFrom, int labelIndexTo, int numPossibleLabels, int maxNumBatches,
-                                       boolean regression, int numThreads, int prefetchSize, boolean collectMetaData, boolean useWorkspaces) {
-
-
-        if (numThreads < 1)
-            numThreads = 1;
-
-        this.workers = numThreads;
-        this.collectMetaData.set(collectMetaData);
-        this.useWorkspaces = useWorkspaces;
-        this.prefetchSize = prefetchSize;
-        this.buffer = new LinkedBlockingQueue<>(prefetchSize);
         this.recordReader = recordReader;
         this.converter = converter;
         this.batchSize = batchSize;
@@ -202,33 +153,75 @@ public class RecordReaderDataSetIterator implements DataSetIterator {
         this.labelIndexTo = labelIndexTo;
         this.numPossibleLabels = numPossibleLabels;
         this.regression = regression;
-
-        log.info("Starting num threads: {}; prefetchBuffer: {}", numThreads, prefetchSize);
-
-        this.executor = (ThreadPoolExecutor) Executors.newFixedThreadPool(numThreads, new ThreadFactory() {
-            @Override
-            public Thread newThread(Runnable r) {
-                Thread t = Executors.defaultThreadFactory().newThread(r);
-
-                // we enforce the same device probably?
-                // TODO: investigate what would be better for perf here
-                //Nd4j.getAffinityManager().attachThreadToDevice(t, Nd4j.getAffinityManager().getDeviceForCurrentThread());
-                t.setDaemon(true);
-                t.setName("RRDSI thread");
-                return t;
-            }
-        });
-
-        // FIXME: fix collectMetaData
-        this.thread = new AsyncPrefetchThread(buffer, recordReader, terminator);
-        Nd4j.getAffinityManager().attachThreadToDevice(this.thread, Nd4j.getAffinityManager().getDeviceForCurrentThread());
-        this.thread.start();
     }
 
 
     @Override
     public DataSet next(int num) {
-        throw new UnsupportedOperationException();
+        if (useCurrent) {
+            useCurrent = false;
+            if (preProcessor != null)
+                preProcessor.preProcess(last);
+            return last;
+        }
+
+        List<DataSet> dataSets = new ArrayList<>();
+        List<RecordMetaData> meta = (collectMetaData ? new ArrayList<RecordMetaData>() : null);
+        for (int i = 0; i < num; i++) {
+            if (!hasNext())
+                break;
+            if (recordReader instanceof SequenceRecordReader) {
+                if (sequenceIter == null || !sequenceIter.hasNext()) {
+                    List<List<Writable>> sequenceRecord = ((SequenceRecordReader) recordReader).sequenceRecord();
+                    sequenceIter = sequenceRecord.iterator();
+                }
+
+                try {
+                    List<Writable> record = sequenceIter.next();
+                    DataSet d = getDataSet(record);
+                    //account for transform process
+                    if (d != null)
+                        dataSets.add(d);
+                }catch(Exception e) {
+                    log.warn("Unable to get dataset ...skipping",e);
+                }
+            } else {
+                if (collectMetaData) {
+                    Record record = recordReader.nextRecord();
+                    DataSet d = getDataSet(record.getRecord());
+                    if(d != null) {
+                        dataSets.add(d);
+                        meta.add(record.getMetaData());
+                    }
+                } else {
+                    try {
+                        List<Writable> record = recordReader.next();
+                        DataSet d = getDataSet(record);
+                        if (d != null)
+                            dataSets.add(d);
+                    }catch(Exception e) {
+                        log.warn("Unable to get dataset ...skipping",e);
+                    }
+                }
+            }
+        }
+        batchNum++;
+
+        if (dataSets.isEmpty()) {
+            return null;
+        }
+
+        DataSet ret = DataSet.merge(dataSets);
+        if (collectMetaData) {
+            ret.setExampleMetaData(meta);
+        }
+        last = ret;
+        if (preProcessor != null)
+            preProcessor.preProcess(ret);
+        //Add label name values to dataset
+        if (recordReader.getLabels() != null)
+            ret.setLabelNames(recordReader.getLabels());
+        return ret;
     }
 
 
@@ -258,13 +251,6 @@ public class RecordReaderDataSetIterator implements DataSetIterator {
             NDArrayWritable writable = (NDArrayWritable) currList.get(0);
             return new DataSet(writable.get(), writable.get());
         }
-
-        if (currList.size() == 2 && currList.get(0) instanceof  NDArrayWritable && currList.get(1) instanceof NDArrayWritable) {
-            NDArrayWritable writableF = (NDArrayWritable) currList.get(0);
-            NDArrayWritable writableL = (NDArrayWritable) currList.get(1);
-            return new DataSet(writableF.get(), writableL.get());
-        }
-
         if (currList.size() == 2 && currList.get(0) instanceof NDArrayWritable) {
             if (!regression) {
                 label = FeatureUtil.toOutcomeVector((int) Double.parseDouble(currList.get(1).toString()),
@@ -391,22 +377,7 @@ public class RecordReaderDataSetIterator implements DataSetIterator {
     @Override
     public void reset() {
         batchNum = 0;
-        thread.shutdown();
-        nextElement = null;
         recordReader.reset();
-        buffer.clear();
-        // TODO: maybe worth shutting down executor as well? or recreate new buffer... or we don't care, Future is gone anyway
-
-        this.thread = new AsyncPrefetchThread(buffer, recordReader, terminator);
-        this.thread.start();
-    }
-
-    public void setCollectMetaData(boolean reallyCollect) {
-        collectMetaData.set(true);
-    }
-
-    public boolean getCollectMetaData() {
-        return getCollectMetaData();
     }
 
     @Override
@@ -432,121 +403,12 @@ public class RecordReaderDataSetIterator implements DataSetIterator {
 
     @Override
     public boolean hasNext() {
-        wasTriggered.compareAndSet(false, true);
-        try {
-            if (nextElement != null && nextElement != terminator) {
-                return true;
-            } else if (nextElement == terminator)
-                return false;
-
-            nextElement = buffer.take();
-
-            if (nextElement == terminator)
-                return false;
-
-            return true;
-        } catch (InterruptedException e) {
-            throw new RuntimeException(e);
-        }
-        //return (recordReader.hasNext() && (maxNumBatches < 0 || batchNum < maxNumBatches));
+        return (recordReader.hasNext() && (maxNumBatches < 0 || batchNum < maxNumBatches));
     }
 
     @Override
     public DataSet next() {
-        if (!wasTriggered.get() && nextElement == null)
-            if (!hasNext())
-                throw new NoSuchElementException("No more records below this line");
-
-
-
-        Future<DataSet> tmp = nextElement;
-        nextElement = null;
-        try {
-            // yes, we're blocking here, but there are chances Future is complete at this moment after first call
-            DataSet ds = tmp.get();
-
-            /*
-            if (ds.getFeatures().isAttached()) {
-                if (Nd4j.getMemoryManager().getCurrentWorkspace() == null) {
-                    ds.detach();
-                } else {
-                    ds.migrate();
-                }
-            }
-            */
-
-            return ds;
-        } catch (Exception e) {
-            throw new RuntimeException(e);
-        }
-
-
-        /*
-        if (useCurrent) {
-            useCurrent = false;
-            if (preProcessor != null)
-                preProcessor.preProcess(last);
-            return last;
-        }
-
-        List<DataSet> dataSets = new ArrayList<>();
-        List<RecordMetaData> meta = (collectMetaData ? new ArrayList<RecordMetaData>() : null);
-        for (int i = 0; i < batchSize; i++) {
-            if (!hasNext())
-                break;
-            if (recordReader instanceof SequenceRecordReader) {
-                if (sequenceIter == null || !sequenceIter.hasNext()) {
-                    List<List<Writable>> sequenceRecord = ((SequenceRecordReader) recordReader).sequenceRecord();
-                    sequenceIter = sequenceRecord.iterator();
-                }
-
-                try {
-                    List<Writable> record = sequenceIter.next();
-                    DataSet d = getDataSet(record);
-                    //account for transform process
-                    if (d != null)
-                        dataSets.add(d);
-                }catch(Exception e) {
-                    log.warn("Unable to get dataset ...skipping",e);
-                }
-            } else {
-                if (collectMetaData) {
-                    Record record = recordReader.nextRecord();
-                    DataSet d = getDataSet(record.getRecord());
-                    if(d != null) {
-                        dataSets.add(d);
-                        meta.add(record.getMetaData());
-                    }
-                } else {
-                    try {
-                        List<Writable> record = recordReader.next();
-                        DataSet d = getDataSet(record);
-                        if (d != null)
-                            dataSets.add(d);
-                    }catch(Exception e) {
-                        log.warn("Unable to get dataset ...skipping",e);
-                    }
-                }
-            }
-        }
-        batchNum++;
-
-        if (dataSets.isEmpty()) {
-            return null;
-        }
-
-        DataSet ret = DataSet.merge(dataSets);
-        if (collectMetaData) {
-            ret.setExampleMetaData(meta);
-        }
-        last = ret;
-        if (preProcessor != null)
-            preProcessor.preProcess(ret);
-        //Add label name values to dataset
-        if (recordReader.getLabels() != null)
-            ret.setLabelNames(recordReader.getLabels());
-        return ret;
-        */
+        return next(batchSize);
     }
 
     @Override
@@ -589,7 +451,7 @@ public class RecordReaderDataSetIterator implements DataSetIterator {
         }
 
         if (dataSets.isEmpty()) {
-           return null;
+            return null;
         }
 
         DataSet ret = DataSet.merge(dataSets);
@@ -600,428 +462,5 @@ public class RecordReaderDataSetIterator implements DataSetIterator {
         if (recordReader.getLabels() != null)
             ret.setLabelNames(recordReader.getLabels());
         return ret;
-    }
-
-
-    protected class AsyncPrefetchThread extends Thread implements Runnable {
-        private BlockingQueue<Future<DataSet>> buffer;
-        private Future<DataSet> terminator;
-        private RecordReader reader;
-        private boolean getMeta;
-        private AtomicBoolean isShutdown = new AtomicBoolean(false);
-        private AtomicBoolean shouldWork = new AtomicBoolean(true);
-        protected RuntimeException exception;
-        protected WorkspaceConfiguration configuration;
-        private String workspaceId;
-
-
-        public AsyncPrefetchThread(@NonNull BlockingQueue<Future<DataSet>> buffer, @NonNull RecordReader reader, @NonNull Future<DataSet> terminator) {
-            this.buffer = buffer;
-            this.terminator = terminator;
-            this.reader = reader;
-            this.workspaceId = "APT_LOOP-" + guid;
-
-
-
-            this.setName("RRDSI prefetch thread");
-            this.setDaemon(true);
-
-            if (reader.batchesSupported()) {
-                configuration = WorkspaceConfiguration.builder()
-                        .overallocationLimit(prefetchSize * Math.max(2, workers))
-                        .minSize(10 * 1024L * 1024L)
-                        .policyMirroring(MirroringPolicy.FULL)
-                        .policySpill(SpillPolicy.EXTERNAL)
-                        .policyLearning(LearningPolicy.FIRST_LOOP)
-                        .policyReset(ResetPolicy.ENDOFBUFFER_REACHED)
-                        .policyAllocation(AllocationPolicy.OVERALLOCATE)
-                        .build();
-            } else {
-                configuration = WorkspaceConfiguration.builder()
-                        .overallocationLimit(batchSize * prefetchSize * Math.max(2, workers))
-                        .minSize(10 * 1024L * 1024L)
-                        .policyMirroring(MirroringPolicy.FULL)
-                        .policySpill(SpillPolicy.EXTERNAL)
-                        .policyLearning(LearningPolicy.FIRST_LOOP)
-                        .policyReset(ResetPolicy.ENDOFBUFFER_REACHED)
-                        .policyAllocation(AllocationPolicy.OVERALLOCATE)
-                        .build();
-            }
-
-            log.info("Workspace overallocation ratio: {}", configuration.getOverallocationLimit());
-        }
-
-
-        @Override
-        public void run() {
-            while (shouldWork.get()) {
-                try {
-                    AtomicLong counterExampes = new AtomicLong(0);
-                    AtomicLong counterOrder = new AtomicLong(0);
-                    getMeta = collectMetaData.get();
-                    boolean limitHit = maxNumBatches > 0 && maxNumBatches < counterOrder.get();
-                    if (!reader.batchesSupported() || getMeta) {
-                        OrderedBatch currentBatch = new OrderedBatch(counterOrder.getAndIncrement());
-                        while (!limitHit && reader.hasNext()) {
-                            if (useWorkspaces) {
-                                try (MemoryWorkspace workspace = Nd4j.getWorkspaceManager().getAndActivateWorkspace(configuration, workspaceId)) {
-                                    if (!getMeta)
-                                        currentBatch.addWritable(reader.next());
-                                    else
-                                        currentBatch.addRecord(reader.nextRecord());
-                                }
-                            } else {
-                                if (!getMeta)
-                                    currentBatch.addWritable(reader.next());
-                                else
-                                    currentBatch.addRecord(reader.nextRecord());
-                            }
-
-                            // if we've built our batch size - we send it to processing
-                            if (currentBatch.size() == batchSize) {
-                                // we should put callable here
-                                currentBatch.commit();
-                                Future<DataSet> future = executor.submit(new DataSetCallable(currentBatch));
-                                getMeta = collectMetaData.get();
-                                buffer.put(future);
-                                currentBatch = new OrderedBatch(counterOrder.getAndIncrement());
-                                limitHit = maxNumBatches > 0 && maxNumBatches < counterOrder.get();
-                            }
-                        }
-                        if (currentBatch.size() > 0 && !limitHit) {
-                            // process last batch
-                            currentBatch.commit();
-                            Future<DataSet> future = executor.submit(new DataSetCallable(currentBatch));
-                            buffer.put(future);
-                        }
-
-                        buffer.put(terminator);
-                    } else {
-                        while (!limitHit && reader.hasNext()) {
-                            if (useWorkspaces) {
-                                try (MemoryWorkspace workspace = Nd4j.getWorkspaceManager().getAndActivateWorkspace(configuration, workspaceId)) {
-                                    long time1 = System.currentTimeMillis();
-                                    List<Writable> batch = reader.next(batchSize);
-                                    DataSet ds = getDataSet(batch);
-                                    long time2 = System.currentTimeMillis();
-
-                                    log.info("Compilation time: {} ms; Footprint: {} bytes", time2 - time1, ds.getMemoryFootprint());
-                                    buffer.put(new DummyFuture(ds));
-                                }
-                            } else {
-                                List<Writable> batch = reader.next(batchSize);
-                                buffer.put(new DummyFuture(getDataSet(batch)));
-                            }
-                            counterOrder.getAndIncrement();
-
-                            limitHit = maxNumBatches > 0 && maxNumBatches < counterOrder.get();
-                        }
-
-                        buffer.put(terminator);
-                    }
-                } catch (InterruptedException e) {
-                    shouldWork.set(false);
-                    isShutdown.set(true);
-                } catch (RuntimeException e) {
-                    e.printStackTrace();
-                    this.exception = e;
-                    isShutdown.set(true);
-                    shouldWork.set(false);
-                }
-            }
-            isShutdown.set(true);
-        }
-
-
-        protected void shutdown() {
-            shouldWork.set(false);
-            thread.interrupt();
-            while (!isShutdown.get()) {
-                LockSupport.parkNanos(100);
-                thread.interrupt();
-            }
-        }
-    }
-
-
-    protected static class OrderedBatch {
-        protected long order;
-        @Getter protected List<List<Writable>> writables = new ArrayList<>();
-        @Getter protected List<Record> records = new ArrayList<>();
-        @Getter protected volatile boolean isRecord = false;
-        protected AtomicInteger counter = new AtomicInteger(0);
-        @Getter protected long timeStart;
-        @Getter protected long timeCommit;
-        @Getter protected long timeCompilation;
-
-        protected OrderedBatch(long order) {
-            this.order = order;
-            timeStart = System.currentTimeMillis();
-        }
-
-        protected void addWritable(List<Writable> writables) {
-            this.writables.add(writables);
-            counter.incrementAndGet();
-        }
-
-        protected void addRecord(Record record) {
-            this.records.add(record);
-            counter.incrementAndGet();
-            isRecord = true;
-        }
-
-        protected int size() {
-            return counter.get();
-        }
-
-
-        protected void commit(){
-            timeCommit = System.currentTimeMillis();
-            timeCompilation = timeCommit - timeStart;
-        }
-    }
-
-
-    protected class DataSetCallable implements Callable<DataSet> {
-        private OrderedBatch batch;
-        private WorkspaceConfiguration configuration;
-        private String workspaceId;
-        protected AtomicBoolean firstLoop = new AtomicBoolean(true);
-
-        public DataSetCallable(OrderedBatch batch) {
-            this.batch = batch;
-
-            configuration = WorkspaceConfiguration.builder()
-                    // FIXME: overalloc limit is wrong here obviously. We should do (divide prefetch size by number of threads) + 1 probably
-                    .overallocationLimit((prefetchSize *2)+1)
-                    .minSize(10 * 1024L * 1024L)
-                    .policyMirroring(MirroringPolicy.FULL)
-                    .policySpill(SpillPolicy.EXTERNAL)
-                    .policyLearning(LearningPolicy.OVER_TIME)
-                    .policyReset(ResetPolicy.ENDOFBUFFER_REACHED)
-                    .policyAllocation(AllocationPolicy.OVERALLOCATE)
-                    .build();
-
-            this.workspaceId = "RRDSI_LOOP-" + guid;
-        }
-
-        /**
-         * This method does OrderedBatch -> DataSet conversion
-         * @return
-         * @throws Exception
-         */
-        @Override
-        public DataSet call() throws Exception {
-
-            DataSet ret = null;
-
-            if (useWorkspaces) {
-                if (Nd4j.getWorkspaceManager().checkIfWorkspaceExists(workspaceId))
-                    firstLoop.set(false);
-
-                // we need to initialize workspace first. so we'll do 2 loops for
-                for (int l = firstLoop.get() ? 0 : 1; l < 2; l++) {
-                    try (MemoryWorkspace workspace = Nd4j.getWorkspaceManager().getAndActivateWorkspace(configuration, workspaceId)) {
-                        // here we create our DataSet
-                        ret = process(batch);
-                    }
-
-                    if (firstLoop.get()) {
-                        Nd4j.getWorkspaceManager().getWorkspaceForCurrentThread(workspaceId).initializeWorkspace();
-                        log.info("Workspace size on initialization: {}", Nd4j.getWorkspaceManager().getWorkspaceForCurrentThread(workspaceId).getCurrentSize());
-                        firstLoop.set(false);
-                    }
-                }
-            } else {
-                ret = process(batch);
-            }
-
-            return ret;
-        }
-
-
-        public DataSet process(OrderedBatch batch) {
-            long timeResume = System.currentTimeMillis();
-            DataSet ret = null;
-            List<DataSet> dataSets = new ArrayList<>();
-            List<RecordMetaData> meta = (batch.isRecord() ? new ArrayList<RecordMetaData>() : null);
-
-            if (batch.isRecord()) {
-                List<Record> records = batch.getRecords();
-                for (int i = 0; i < records.size(); i++) {
-                    DataSet d = getDataSet(records.get(i).getRecord());
-                    dataSets.add(d);
-                    meta.add(records.get(i).getMetaData());
-                }
-            } else {
-                List<List<Writable>> writables = batch.getWritables();
-                for (int i = 0; i < writables.size(); i++) {
-                    DataSet d = getDataSet(writables.get(i));
-                    dataSets.add(d);
-                }
-            }
-
-            long timePull = System.currentTimeMillis();
-
-            // should NOT ever happen
-            if (dataSets.isEmpty()) {
-                return null;
-            }
-
-            ret = DataSet.merge(dataSets);
-
-            if (batch.isRecord()) {
-                ret.setExampleMetaData(meta);
-            }
-
-            if (preProcessor != null)
-                preProcessor.preProcess(ret);
-
-            //Add label name values to dataset
-            if (recordReader.getLabels() != null)
-                ret.setLabelNames(recordReader.getLabels());
-
-            long timeMerge = System.currentTimeMillis();
-
-            log.info("Compilation: {} ms; Pull: {} ms; Merge: {} ms; Travel: {} ms;", batch.getTimeCompilation(), timePull - timeResume, timeMerge - timePull, timeMerge - batch.getTimeStart());
-
-            return ret;
-        }
-    }
-
-
-    protected static class DummyFuture implements Future<DataSet> {
-        DataSet dataSet;
-
-        public DummyFuture() {
-
-        }
-
-        public DummyFuture(DataSet ds) {
-            this.dataSet = ds;
-        }
-
-        @Override
-        public boolean cancel(boolean mayInterruptIfRunning) {
-            return false;
-        }
-
-        @Override
-        public boolean isCancelled() {
-            return false;
-        }
-
-        @Override
-        public boolean isDone() {
-            return true;
-        }
-
-        @Override
-        public DataSet get() throws InterruptedException, ExecutionException {
-            return dataSet;
-        }
-
-        @Override
-        public DataSet get(long timeout, TimeUnit unit) throws InterruptedException, ExecutionException, TimeoutException {
-            return dataSet;
-        }
-    }
-
-    public static class Builder {
-        protected RecordReader recordReader;
-        protected WritableConverter converter = new SelfWritableConverter();
-        protected int batchSize = 10;
-        protected int maxNumBatches = -1;
-        protected int labelIndex = -1;
-        protected int labelIndexTo = -1;
-        protected int numPossibleLabels = -1;
-        protected boolean regression = false;
-        protected DataSetPreProcessor preProcessor;
-        private boolean collectMetaData = false;
-        private int prefetchSize = 2;
-        private int workers = 1;
-        private boolean useWorkspaces = true;
-
-        public Builder(@NonNull RecordReader reader) {
-            this.recordReader = reader;
-            if (reader.getLabels()!=null)
-                numPossibleLabels = reader.getLabels().size();
-        }
-
-        public Builder setWritableConverter(@NonNull WritableConverter converter) {
-            this.converter = converter;
-            return this;
-        }
-
-        public Builder setBatchSize(int batchSize) {
-            if (batchSize < 1)
-                throw new DL4JInvalidConfigException("batchSize can't be negative value");
-
-            this.batchSize = batchSize;
-            return this;
-        }
-
-        public Builder setMaxNumberOfDataSets(int size) {
-            this.maxNumBatches = size;
-            return this;
-        }
-
-        public Builder setRegressionMode(boolean reallySet) {
-            this.regression = reallySet;
-            return this;
-        }
-
-        public Builder setLabelIndex(int idx) {
-            this.labelIndex = idx;
-            return this;
-        }
-
-        public Builder setLabelIndexTo(int idx) {
-            this.labelIndexTo = idx;
-            return this;
-        }
-
-
-        public Builder setNumberOfPossibleLabels(int number) {
-            this.numPossibleLabels = number;
-            return this;
-        }
-
-        public Builder collectMetaData(boolean reallyCollect) {
-            this.collectMetaData = reallyCollect;
-            return this;
-        }
-
-        public Builder useWorkspaces(boolean reallyUse) {
-            this.useWorkspaces = reallyUse;
-            return this;
-        }
-
-        public Builder setDataSetPreProcessor(DataSetPreProcessor preProcessor) {
-            this.preProcessor = preProcessor;
-            return this;
-        }
-
-        public Builder prefetchBufferSize(int size) {
-            if (size < 2)
-                size = 2;
-
-            this.prefetchSize = size;
-            return this;
-        }
-
-        public Builder numberOfWorkers(int workers) {
-            if (workers < 1)
-                workers = 1;
-
-            this.workers = workers;
-            return this;
-        }
-
-
-        public RecordReaderDataSetIterator build() {
-            RecordReaderDataSetIterator rrdsi = new RecordReaderDataSetIterator(recordReader, converter, batchSize, labelIndex, labelIndexTo, numPossibleLabels, maxNumBatches, regression, workers, prefetchSize, collectMetaData, useWorkspaces);
-            return rrdsi;
-        }
     }
 }
